Update pod2html
> [...]
> # Lost interest?  It's so bad I've lost apathy.

You too, eh?

However, here's a patch which brings the _91 version up to date.

YES, it still slurps chunks.      ( No comment )

YES, it still uses XMP.           (Nothing else works right)

YES, it still is hard to follow.  (But oh, the joy of figuring it out!)

YES, it works on all platforms with configuration, including Macs...

I had pretty much given up on this, just telling people to grab
the latest from my CPAN dir, but I get alot of mail with fixes
against 1.15, and this is version is vastly better...  Please
try it, and apply to 5.004, if it's not too late.

I will (respectfully) ignore any whining about the inappropriateness
of some of the things I had to do in here.  Also any comments 
about other renditions, I've tried them all, and this one still 
excels; besides working on all platforms.  I'm in the middle of a 
rather difficult project, and I took the day off today to get this 
together.  

p5p-msgid: 199703030025.QAA08106@ducks