[differences between cumulative patch application and perl5.004_01]
[editor's note: The changes between this and 5.004 were processed from
the m1t2 release, which was a bad idea as it was the _01 release which
had the final corrected attributions.  The differences between the
various m*t* releases do that; I considered it most valuable just to
look at the _NN releases.  Many patches have been separated out and/or
applied from the p5p archives nonetheless.]