make Pod-Html's _unixify normalize Win32 drive letters
Okay, look, this is kind of stupid and horrible, but it should
stop smoke failures and actually address some forms of failed path
comparisons in Pod::Html.  In reality, we should be doing better path
comparisons than checking substr and eq, but that's not going to get
properly overhauled at this late date.

In the meantime, this should fix Win32 smokers with forcibly lc()-ed
cwds without breaking anybodye else, right?  Right!